refactor(synapsys): extract Phase 1 consume logic to keep cortex-recall under 400 lines
Extract consume-cache.js with Phase 1 consumption and baseline deferral logic from cortex-recall.js. No behavior change — extraction to meet size and complexity gates.
